split rules::req_ref into three files
